requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
opencv-python-headless
scikit-image
numpy
python-multipart
python-dotenv
azure-storage-blob
aiofiles
